refactor(hal): use uintptr_t for hardware address enums, rename smc_generated to smc
Replace uint32_t with uintptr_t as the underlying type for all C23 typed enums
that hold RX72N hardware peripheral base addresses across 30 files. uintptr_t is
the semantically correct type for storing pointer-sized addresses, whereas uint32_t
expresses only integer width. On RX72N (32-bit) both are the same width, so this
is a correctness and documentation improvement with zero runtime cost.

Also renames src/boot/smc_generated/ and src/boot/inc/smc_generated/ to smc/ to
align with STAR coding standards (dropping the redundant _generated suffix).

Files changed:
- 28x rx_hal/inc/rx72n_*_regs.h: address enums uint32_t -> uintptr_t
- tests/mocks/rx72n_system_regs.h: same fix for mock address enums
- src/boot/smc/lowlvl.c: E1 debug port address enum uint32_t -> uintptr_t
- src/boot/smc_generated/ -> src/boot/smc/ (renamed, 5 source files)
- src/boot/inc/smc_generated/ -> src/boot/inc/smc/ (renamed, 7 header files)
- CMakeLists.txt: update include path from inc/smc_generated to inc/smc

+++ star-rx72n-firmware/libs/rx_hal/inc/rx72n_flash_regs.h
@@ -118,7 +118,7 @@ extern "C" {
  *
  * @since Version 1.0.0
  */
-typedef enum : uint32_t {
+typedef enum : uintptr_t {
   k_rom_cache_base_addr = 0x00081000, /**< ROM cache register base */
   k_romce_addr          = 0x00081000, /**< ROM Cache Enable Register */
   k_romciv_addr         = 0x00081004, /**< ROM Cache Invalidate Register */
@@ -187,7 +187,7 @@ static inline volatile uint16_t* romciv_reg(void)
  *
  * @since Version 1.0.0
  */
-typedef enum : uint32_t {
+typedef enum : uintptr_t {
   k_fwepror_addr = 0x0008C296, /**< Flash P/E Protect Register */
 } rx_fwepror_addresses_t;
